Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/web-share/share-simple-manual.https-expected.txt

Instructions:

Click the Share button.
Choose a valid share target.
Pass the test iff the target app received:

title = "the title"
text = "the message"
url = "https://www.example.com/some/path?some_query#some_fragment"


Harness Error (TIMEOUT), message = null
